backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
google-generativeai
pillow
python-dotenv

flask
flask-cors
